fix(modal): Remove remote scoping to allow partial build with ivy
