+++ src/modal/index.ts
@@ -10,6 +10,7 @@ export { ModalFooter } from "./modal-footer.component";
 export { ModalHeader } from "./modal-header.component";
 export { Modal } from "./modal.component";
 export { ModalModule } from "./modal.module";
+export { BaseModalService } from "./base-modal.service";
 export { ModalService } from "./modal.service";
 export { Overlay } from "./overlay.component";
 export { ModalContent } from "./modal-content.directive";
