+++ src/tsconfig.lib.json
@@ -1,11 +1,11 @@
 {
-  "extends": "./../tsconfig.json",
-  "compilerOptions": {
-    "declarationMap": false,
-	  "rootDir": "."
-  },
-  "angularCompilerOptions": {
-    "enableIvy": true,
-    "compliationMode": "partial"
-  }
+	"extends": "./../tsconfig.json",
+	"compilerOptions": {
+		"declarationMap": false,
+		"rootDir": "."
+	},
+	"angularCompilerOptions": {
+		"enableIvy": true,
+		"compliationMode": "partial"
+	}
 }
